Add workaround for missing /dev/tun with device plugins
In KubeVirt >= 0.7.0 device plugins are used to access /dev/kvm (if not
in emulation mode) and /dev/tun. Minishift and Minikube versions based
on Kubernetes 1.10 however do not automatically expose these.

To combat this a workarounds.sh script is added. This is mounted into
the minishift VM and executed before deploying Kubernetes on the
cluster. The script enables the feature gate for device plugins, and
exposes the relevant devices to the kubelet thus ensuring that they will
be available when the virt-handler comes looking for them.